requirements.txt
aiohttp==3.6.2
alabaster==0.7.12
alembic==1.4.2
appdirs==1.4.4
astroid==2.4.1
async-timeout==3.0.1
atpublic==1.0
attrs==19.3.0
autopep8==1.5.3
Babel==2.9.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.1.5
bokeh==2.1.1
cdflib==0.3.18
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.3
colorcet==2.0.2
configobj==5.0.6
controldiffeq==0.0.1
cryptography==2.9.2
cycler==0.10.0
dask==2.19.0
databricks-cli==0.11.0
datashader==0.11.0
datashape==0.5.2
debugpy==1.0.0b10
decorator==4.4.2
defusedxml==0.6.0
distributed==2.19.0
distro==1.5.0
docker==4.2.1
docutils==0.16
dpath==2.0.1
dvc==0.94.0
entrypoints==0.3
fabric==2.5.0
Faker==4.1.0
filelock==3.0.12
flake8==3.8.1
Flask==1.1.2
flatten-json==0.1.7
flufl.lock==3.2
fsspec==0.7.4
funcy==1.14
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
google==2.0.3
gorilla==0.3.0
grandalf==0.6
grpcio==1.29.0
gunicorn==20.0.4
HeapDict==1.0.1
holoviews==1.13.3
humanize==2.4.0
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.1
inflect==3.0.2
invoke==1.4.1
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
json5==0.9.5
jsonpath-ng==1.5.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
jupyterlab-widgets==1.0.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
locket==0.2.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mlflow==1.8.0
msgpack==0.6.2
multidict==4.7.6
multipledispatch==0.6.0
nanotime==0.5.2
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.3
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.19.5
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
panel==0.9.7
param==1.9.3
paramiko==2.7.1
parso==0.7.0
partd==1.1.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.8.2
ply==3.11
prometheus-client==0.8.0
prometheus-flask-exporter==0.13.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
py-spy==0.3.3
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pyct==0.4.6
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.6.1
pygtrie==2.3.2
pylint==2.5.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.16.0
pysftp==0.2.9
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pyviz-comms==0.7.5
PyYAML==5.3.1
pyzmq==19.0.1
querystring-parser==1.2.4
ray==0.8.5
redis==3.4.1
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
shortuuid==1.0.1
simplejson==3.17.0
six==1.15.0
smmap==3.0.4
snowballstemmer==2.1.0
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx==3.1.1
sphinx-autoapi==1.4.0
sphinx-autodoc-typehints==1.11.0
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.13
sqlparse==0.3.1
tables==3.6.1
tabulate==0.8.7
tblib==1.6.0
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
texttable==1.6.2
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
torch==1.5.0
torchdiffeq==0.0.1
torchsummary==1.5.1
torchvision==0.6.0
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
treelib==1.6.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
Unidecode==1.2.0
urllib3==1.25.9
voluptuous==0.11.7
wcwidth==0.2.3
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.15.1
yarl==1.4.2
zc.lockfile==2.0
zict==2.0.0
zipp==3.1.0
git+https://github.com/patrick-kidger/NeuralCDE.git
